+++ tests/test_interference.py
@@ -3,13 +3,15 @@ import xmostest
 from i2c_master_checker import I2CMasterChecker
 import os
 
-def runtest():
+def do_master_test(arch, stop):
     resources = xmostest.request_resource("xsim")
 
     speed = 100
-    build_config = "interfere"
 
-    binary = 'i2c_master_async_test/bin/%(config)s/i2c_master_async_test_%(config)s.xe' % {'config':build_config}
+    binary = 'i2c_master_async_test/bin/interfere_%(arch)s_%(stop)s/i2c_master_async_test_interfere_%(arch)s_%(stop)s.xe' % {
+     'arch' : arch,
+     'stop' : stop,
+    }
 
     checker = I2CMasterChecker("tile[0]:XS1_PORT_1A",
                                "tile[0]:XS1_PORT_1B",
@@ -21,13 +23,21 @@ def runtest():
                                              True, True, True, False,
                                              True, False])
 
-    tester = xmostest.ComparisonTester(open('master_test.expect'),
+    tester = xmostest.ComparisonTester(open('master_test_%s.expect' % stop),
                                      'lib_i2c', 'i2c_master_sim_tests',
-                                     'async_interference_test', {'speed':str(speed)},
+                                     'async_interference_test',
+                                     {'speed':speed, 'arch' : arch, 'stop' : stop},
                                      regexp=True)
+
+    sim_args = ['--weak-external-drive']
 
     xmostest.run_on_simulator(resources['xsim'], binary,
                               simthreads = [checker],
-                              simargs=['--weak-external-drive'],
+                              simargs=sim_args,
                               suppress_multidrive_messages = True,
                               tester = tester)
+
+def runtest():
+  for arch in ['xs1', 'xs2']:
+    for stop in ['stop', 'no_stop']:
+      do_master_test(arch, stop)