+++ tests/test_reg_ops.py
@@ -4,10 +4,12 @@ from i2c_master_checker import I2CMasterChecker
 import os
 
 
-def runtest():
+def do_test(arch):
     resources = xmostest.request_resource("xsim")
 
-    binary = 'i2c_master_reg_test/bin/i2c_master_reg_test.xe'
+    binary = 'i2c_master_reg_test/bin/%(arch)s/i2c_master_reg_test_%(arch)s.xe' % {
+      'arch' : arch
+    }
 
     checker = I2CMasterChecker("tile[0]:XS1_PORT_1A",
                                "tile[0]:XS1_PORT_1B",
@@ -25,11 +27,15 @@ def runtest():
     tester = xmostest.ComparisonTester(open('reg_test.expect'),
                                      'lib_i2c', 'i2c_master_sim_tests',
                                      'reg_ops_test',
+                                     {'arch' : arch},
                                      regexp=True)
 
     xmostest.run_on_simulator(resources['xsim'], binary,
-                              simthreads = [checker],
+                              simthreads=[checker],
                               simargs=['--weak-external-drive'],
-                              suppress_multidrive_messages = True,
-                              tester = tester)
+                              suppress_multidrive_messages=True,
+                              tester=tester)
 
+def runtest():
+  for arch in ['xs1', 'xs2']:
+    do_test(arch)